fix: page fault handler
The IDT entry number 14 (Page Fault) was missing. Copy pasting is
bad, even with vim macros.

+++ kernel/src/cpu/gdt.c
@@ -8,13 +8,14 @@ static gdt_entry_ptr_t gdt_ptr;
 void init_gdt() {
 	gdt_ptr.limit = (sizeof(gdt_entry_t)*5) - 1;
 	gdt_ptr.base = (uint32_t) &gdt_entries;
+
 	gdt_set_entry(0, 0, 0, 0, 0);                // Null segment
 	gdt_set_entry(1, 0, 0xFFFFFFFF, 0x9A, 0xCF); // Code segment
 	gdt_set_entry(2, 0, 0xFFFFFFFF, 0x92, 0xCF); // Data segment
 	gdt_set_entry(3, 0, 0xFFFFFFFF, 0xFA, 0xCF); // User mode code segment
 	gdt_set_entry(4, 0, 0xFFFFFFFF, 0xF2, 0xCF); // User mode data segment
 
-	gdt_load((uint32_t) &gdt_ptr);
+	gdt_load((uintptr_t) &gdt_ptr);
 }
 
 void gdt_set_entry(uint32_t num, uint32_t base, uint32_t limit, uint8_t access, uint8_t granularity) {
